Removes compilation warning

+++ src/websocket.cpp
@@ -59,13 +59,13 @@ long long webSocket::frameSize(const string &data
 	mask_size = 0;
 	payload_size = 0;
 	const unsigned char *rawdata = (const unsigned char *)data.data();
-	if (data.size() < control_size + mask_size + payload_size) {
+	if (data.size() < (unsigned long long) (control_size + mask_size + payload_size)) {
 		return -1;
 	}
 	//Is masked frame (must be musked)
 	mask_size = (rawdata[1] & 0x80) ? 4:0;
 	payload_size = rawdata[1] & 0x7f;
-	if( data.size() < control_size + mask_size + payload_size) {
+	if( data.size() < (unsigned long long) (control_size + mask_size + payload_size)) {
 		return -1;
 	}
 	if (payload_size == 126) {
